+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.38.0 - TBD
+## 0.38.0 - 2024-07-23
 
 This release adds a new feature to the `Live` client for automatically reconnecting when an unexpected disconnection occurs.